Remove int parsing

+++ lumen/app/Jobs/ImportEvent.php
@@ -36,7 +36,7 @@ class ImportEvent extends Job
 
         if (empty($event)) {
             $venueStorage = $client->test->venues;
-            $venueId = (int) data_get($toImport, 'metadata.venue_id');
+            $venueId = data_get($toImport, 'metadata.venue_id');
 
             $venue = $venueStorage->findOne([
                 'external_id' => $venueId
